Move album canonical into service layer
This removes the last DB-backed utility seam from utils and enforces the canonical deps.db contract for album canonical operations, while keeping route wiring and compatibility tests aligned.
